Revert "Login functionality"
