+++ src/actions/index.js
@@ -2,8 +2,3 @@ export const getMovies = movies => ({
   type: 'GET_MOVIES',
   movies
 });
-
-export const updateUser = user => ({
-  type: 'UPDATE_USER',
-  user
-});
