+++ src/reducers/index.js
@@ -1,8 +1,6 @@
 import { combineReducers } from 'redux';
 import { movies } from './moviesReducer'
-import { user } from './userReducer'
 
 export const rootReducer = combineReducers({
-  movies,
-  user
+  movies
 });
