--- src/reducers/userReducer.js
@@ -1,8 +1,0 @@
-export const user = (state = {}, action) => {
-  switch(action.type) {
-    case 'UPDATE_USER':
-      return action.user
-    default:
-      return state;
-  }
-}
